+++ ios/Classes/Chats/FlyChatMethods.swift
@@ -421,7 +421,7 @@ let ISEXPORT = true
             if let unformattedPhoneNumber = try? phoneNumberUtil.format(phoneNumber, numberFormat: .E164).replacingOccurrences(of: "+", with: "") {
                 do{
                     try result(FlyUtils.getJid(from: unformattedPhoneNumber))
-                }catch let jidError{
+                }catch _{
                     result(FlutterError(code: FLErrorCode.INVALID_DATA,message: FLErrorMessage.JID_FETCH_FAILED,details: nil))
                 }
             }
@@ -817,7 +817,7 @@ let ISEXPORT = true
                 if((data.getData() as? [ProfileDetails])?.count != 0){
                     userData = (list?.toJson())!
                 }
-                result(userData)
+                result("{\"data\":" + userData + "}")
             } else{
                 if case let .unexpected(message, code) = flyError {
                     if code == ErrorCode.NO_NETWORK{
@@ -1347,7 +1347,7 @@ let ISEXPORT = true
         
         let savedMessage = FlyMessenger.getUnsentMessageOf(id: userjid)
         let getUnsentMessageJSON = "{\"textContent\" : \"\(savedMessage.textContent)\",\"mentionedUsers\": " + (savedMessage.mentionedUsers.toJson() ?? "[]") + "}"
-        print("savedMessage toJson : \(savedMessage.toJson())")
+        print("savedMessage toJson : \(String(describing: savedMessage.toJson()))")
         print("savedMessage : \(getUnsentMessageJSON)")
         result(getUnsentMessageJSON)
         
@@ -1942,6 +1942,18 @@ let ISEXPORT = true
         let muteStatus = args["mute_status"] as? Bool ?? false
         ChatManager.updateChatMuteStatus(jid: userJID, muteStatus: muteStatus)
     }
+    
+    func updateChatMuteStatusList(call: FlutterMethodCall, result: @escaping FlutterResult){
+        
+        let args = call.arguments as! Dictionary<String, Any>
+        
+        let userJidList = args["jidList"] as? [String] ?? []
+        let muteStatus = args["mute_status"] as? Bool ?? false
+        
+        ChatManager.updateChatMuteStatus(jidList: userJidList, mute: muteStatus)
+        
+    }
+        
     func sendTypingStatus(call: FlutterMethodCall, result: @escaping FlutterResult){
         
         let args = call.arguments as! Dictionary<String, Any>
@@ -2812,7 +2824,11 @@ let ISEXPORT = true
                 }else{
                     print("\(Constants.tag) Next Message List last message id is not setting as the list is empty")
                 }
-                
+                if(self.firstMessageID.isEmpty){
+                    self.firstMessageID = messageList?.first?.messageId ?? emptyString()
+                    self.setFirstMessage()
+                }
+
                 if let chatJson = messageList.toJson() {
                     print("\(Constants.tag) Next Message List \(chatJson)")
                     result(chatJson)
@@ -3051,6 +3067,10 @@ let ISEXPORT = true
         ChatManager.clearChat(toJid: userJid, chatType: chatType!, clearChatExceptStarred: clearExceptStarred) { (isSuccess, flyError, resultDict) in
             
             if(isSuccess){
+                self.lastMessageID = emptyString()
+                self.setLastMessage()
+                self.firstMessageID = emptyString()
+                self.setFirstMessage()
                 result(true)
             }else{
                 if case let .invalid_data(message, _) = flyError {
@@ -3859,7 +3879,14 @@ let ISEXPORT = true
         let replyMessageID = args["replyMessageId"] as? String
         let topicId = args["topicId"] as? String
         let mentionedUsersIds = args["mentionedUsersIds"] as? [String] ?? []
-        
+        let metaData = args["metaData"] as? [[String: Any]] ?? []
+        print("\(String(describing: messageType)) MetaData \(String(describing: metaData))")
+        var metaDataArray : [MessageMetaData] = []
+        for data in metaData {
+            let obj = MessageMetaData(key: data["key"] as? String ?? "", value: data["value"] as? String ?? "")
+            metaDataArray.append(obj)
+        }
+
         if let sendingMessageType = FlyMessageType.fromString(messageType ?? "") {
             print("\(Constants.tag) sendMessage -> Messsage Type \(sendingMessageType)")
             switch sendingMessageType {
@@ -3870,13 +3897,6 @@ let ISEXPORT = true
                 (AppUtils.shared.getValueForKey(dictionary: args["textMessage"] as? Dictionary<String, Any>, key: "messageText") as? String ?? "").trimmingCharacters(in: .whitespacesAndNewlines)
                 messageParams.replyMessageId = replyMessageID
                 messageParams.mentionedUsersIds = mentionedUsersIds
-                let metaData = args["metaData"] as? [[String: Any]] ?? []
-                print("Text MetaData \(String(describing: metaData))")
-                var metaDataArray : [MessageMetaData] = []
-                for data in metaData {
-                    let obj = MessageMetaData(key: data["key"] as? String ?? "", value: data["value"] as? String ?? "")
-                    metaDataArray.append(obj)
-                }
                 messageParams.metaData = metaDataArray
                 messageParams.topicID = topicId
                 
@@ -3892,14 +3912,6 @@ let ISEXPORT = true
                 let fileThumbImageArg = AppUtils.shared.getValueForKey(dictionary: fileDictArg, key: "thumbImage") as? String ?? ""
                 let fileNameArg = AppUtils.shared.getValueForKey(dictionary: fileDictArg, key: "fileName") as? String ?? ""
                 let fileCaptionArg = AppUtils.shared.getValueForKey(dictionary: fileDictArg, key: "caption") as? String ?? ""
-                
-                let metaData = args["metaData"] as? [[String: Any]] ?? []
-                print("Image MetaData \(String(describing: metaData))")
-                var metaDataArray : [MessageMetaData] = []
-                for data in metaData {
-                    let obj = MessageMetaData(key: data["key"] as? String ?? "", value: data["value"] as? String ?? "")
-                    metaDataArray.append(obj)
-                }
 
                 let imagefileUrl = URL(fileURLWithPath: filePathArg)
                 
@@ -3920,7 +3932,7 @@ let ISEXPORT = true
                         
                         let mediaParams = FileMessageParams(fileUrl: localFilePath!, fileName: fileNameArg == "" ? fileName : fileNameArg,  caption : fileCaptionArg, fileSize: fileSize, duration: 0.0, thumbImage: fileThumbImageArg == "" ? MediaUtils.convertImageToBase64String(img: selectedImage!) : fileThumbImageArg, fileKey: fileKey)
                         
-                        let imageFileMessage = FileMessage(toId: receiverJID!, messageType: .image, fileMessage : mediaParams, replyMessageId : replyMessageID, mentionedUsersIds: mentionedUsersIds, metaData: metaDataArray)
+                        let imageFileMessage = FileMessage(toId: receiverJID!, messageType: .image, fileMessage : mediaParams, replyMessageId : replyMessageID, mentionedUsersIds: mentionedUsersIds, metaData: metaDataArray, topicID: topicId)
                         
                         self.sendImage(imageMessageParams: imageFileMessage, call: call, result: result)
                         
@@ -3939,14 +3951,6 @@ let ISEXPORT = true
                 _ = AppUtils.shared.getValueForKey(dictionary: fileDictArg, key: "thumbImage") as? String ?? ""
                 _ = AppUtils.shared.getValueForKey(dictionary: fileDictArg, key: "fileName") as? String ?? ""
                 let fileCaptionArg = AppUtils.shared.getValueForKey(dictionary: fileDictArg, key: "caption") as? String ?? ""
-                
-                let metaData = args["metaData"] as? [[String: Any]] ?? []
-                print("Image MetaData \(String(describing: metaData))")
-                var metaDataArray : [MessageMetaData] = []
-                for data in metaData {
-                    let obj = MessageMetaData(key: data["key"] as? String ?? "", value: data["value"] as? String ?? "")
-                    metaDataArray.append(obj)
-                }
 
                 let videoFileUrl = URL(fileURLWithPath: filePathArg)
                 
@@ -3967,7 +3971,7 @@ let ISEXPORT = true
                     if let compressedURL = url,  isSuccess{
                         
                         let mediaParams = FileMessageParams(fileUrl: compressedURL, fileName: fileName, caption: fileCaptionArg, fileSize: fileSize, duration: duration, thumbImage: base64Img, fileKey: fileKey)
-                        let videoFileMessage = FileMessage(toId: receiverJID!, messageType: .video, fileMessage : mediaParams, replyMessageId: replyMessageID, mentionedUsersIds: mentionedUsersIds, metaData: metaDataArray)
+                        let videoFileMessage = FileMessage(toId: receiverJID!, messageType: .video, fileMessage : mediaParams, replyMessageId: replyMessageID, mentionedUsersIds: mentionedUsersIds, metaData: metaDataArray, topicID: topicId)
                         
                         self.sendVideo(videoMessageParams: videoFileMessage, call: call, result: result)
                         
@@ -3987,21 +3991,14 @@ let ISEXPORT = true
                 _ = AppUtils.shared.getValueForKey(dictionary: fileDictArg, key: "thumbImage") as? String ?? ""
                 _ = AppUtils.shared.getValueForKey(dictionary: fileDictArg, key: "fileName") as? String ?? ""
                 _ = AppUtils.shared.getValueForKey(dictionary: fileDictArg, key: "caption") as? String ?? ""
-                
-                let metaData = args["metaData"] as? [[String: Any]] ?? []
-                print("Image MetaData \(String(describing: metaData))")
-                var metaDataArray : [MessageMetaData] = []
-                for data in metaData {
-                    let obj = MessageMetaData(key: data["key"] as? String ?? "", value: data["value"] as? String ?? "")
-                    metaDataArray.append(obj)
-                }
+
 
                 let audiofileUrl = URL(fileURLWithPath: filePathArg)
                 
                 MediaUtils.processAudioFile(url: audiofileUrl) { isSuccess, fileName ,localPath, fileSize, duration, fileKey, errorMessage  in
                     if let localPathURL = localPath, isSuccess{
                         let audioParams = FileMessageParams (fileUrl: localPathURL, fileName: fileName,fileSize: fileSize, duration: duration, fileKey: fileKey)
-                        let audioFileMessage = FileMessage(toId: receiverJID ?? emptyString(), messageType: sendingMessageType == .AUDIO_RECORDED ? .audioRecorded : .audio, fileMessage : audioParams, replyMessageId: replyMessageID ?? emptyString(), metaData: metaDataArray)
+                        let audioFileMessage = FileMessage(toId: receiverJID ?? emptyString(), messageType: sendingMessageType == .AUDIO_RECORDED ? .audioRecorded : .audio, fileMessage : audioParams, replyMessageId: replyMessageID ?? emptyString(), metaData: metaDataArray, topicID: topicId)
                         self.sendAudio(audioMessageParams: audioFileMessage, call: call, result: result)
                         
                     } else {
@@ -4017,15 +4014,7 @@ let ISEXPORT = true
                 let contactName = AppUtils.shared.getValueForKey(dictionary: contactDict, key: "name") as? String ?? ""
                 let contactNumbers = AppUtils.shared.getValueForKey(dictionary: contactDict, key: "numbers") as? [String] ?? []
                 
-                let metaData = args["metaData"] as? [[String: Any]] ?? []
-                print("Image MetaData \(String(describing: metaData))")
-                var metaDataArray : [MessageMetaData] = []
-                for data in metaData {
-                    let obj = MessageMetaData(key: data["key"] as? String ?? "", value: data["value"] as? String ?? "")
-                    metaDataArray.append(obj)
-                }
-
-                let contactMessageParams = FileMessage(toId: receiverJID!, messageType: .contact, contactMessage: ContactMessageParams(name: contactName, numbers: contactNumbers), replyMessageId: replyMessageID ?? emptyString(), metaData: metaDataArray)
+                let contactMessageParams = FileMessage(toId: receiverJID!, messageType: .contact, contactMessage: ContactMessageParams(name: contactName, numbers: contactNumbers), replyMessageId: replyMessageID ?? emptyString(), metaData: metaDataArray, topicID: topicId)
                 
                 sendContact(contactMessageParams: contactMessageParams, call: call, result: result)
                 break;
@@ -4033,14 +4022,6 @@ let ISEXPORT = true
             case .DOCUMENT:
                 let fileDictArg = args["fileMessage"] as? Dictionary<String, Any>
                 let filePathArg = AppUtils.shared.getValueForKey(dictionary: fileDictArg, key: "file") as? String ?? ""
-                
-                let metaData = args["metaData"] as? [[String: Any]] ?? []
-                print("Image MetaData \(String(describing: metaData))")
-                var metaDataArray : [MessageMetaData] = []
-                for data in metaData {
-                    let obj = MessageMetaData(key: data["key"] as? String ?? "", value: data["value"] as? String ?? "")
-                    metaDataArray.append(obj)
-                }
 
                 let documentFileUrl = URL(fileURLWithPath: filePathArg)
                 
@@ -4049,7 +4030,7 @@ let ISEXPORT = true
                     if let localPathURL = localPath, isSuccess {
                         
                         let documentParams = FileMessageParams(fileUrl: localPathURL, fileName: fileName)
-                        let documentMsg = FileMessage(toId: receiverJID!, messageType: .document, fileMessage: documentParams, replyMessageId: replyMessageID ?? emptyString(), metaData: metaDataArray)
+                        let documentMsg = FileMessage(toId: receiverJID!, messageType: .document, fileMessage: documentParams, replyMessageId: replyMessageID ?? emptyString(), metaData: metaDataArray, topicID: topicId)
                         
                         self.sendDocument(documentMessageParams: documentMsg, call: call, result: result)
                         
@@ -4067,21 +4048,24 @@ let ISEXPORT = true
                 let locationDict = args["locationMessage"] as? Dictionary<String, Any>
                 let latitude = AppUtils.shared.getValueForKey(dictionary: locationDict, key: "latitude") as? Double ?? 0.0
                 let longitude = AppUtils.shared.getValueForKey(dictionary: locationDict, key: "longitude") as? Double ?? 0.0
-                
-                let metaData = args["metaData"] as? [[String: Any]] ?? []
-                print("Image MetaData \(String(describing: metaData))")
-                var metaDataArray : [MessageMetaData] = []
-                for data in metaData {
-                    let obj = MessageMetaData(key: data["key"] as? String ?? "", value: data["value"] as? String ?? "")
-                    metaDataArray.append(obj)
-                }
 
-                let locationMessageParams = FileMessage(toId: receiverJID!, messageType: .location, locationMessage: LocationMessageParams(latitude: latitude, longitude: longitude), replyMessageId: replyMessageID ?? emptyString(), metaData: metaDataArray)
+                let locationMessageParams = FileMessage(toId: receiverJID!, messageType: .location, locationMessage: LocationMessageParams(latitude: latitude, longitude: longitude), replyMessageId: replyMessageID ?? emptyString(), metaData: metaDataArray, topicID: topicId)
                 
                 sendLocation(locationMessageParams: locationMessageParams, call: call, result: result)
                 
                 break;
                 
+            case .MEET:
+                let meetDict = args["meetMessage"] as? Dictionary<String, Any>
+                let title = AppUtils.shared.getValueForKey(dictionary: meetDict, key: "title") as? String ?? ""
+                let link = AppUtils.shared.getValueForKey(dictionary: meetDict, key: "link") as? String ?? ""
+                let scheduledDateTime = AppUtils.shared.getValueForKey(dictionary: meetDict, key: "scheduledDateTime") as? Int ?? 0
+
+                let meetMessageParams = MeetMessage(toId: receiverJID!, title: title, link: link, scheduledDateTime: scheduledDateTime,replyMessageId: replyMessageID ?? emptyString(),mentionedUsersIds: mentionedUsersIds,metaData: metaDataArray,topicID: topicId)
+
+                sendMeetMessage(meetMessageParams: meetMessageParams, call: call, result: result)
+
+                break;
             default:
                 print("sendMessage -> Messsage Type goes to Default")
                 break;
@@ -4307,6 +4291,37 @@ let ISEXPORT = true
             }
         }
     }
+
+    private func sendMeetMessage(meetMessageParams: MeetMessage, call : FlutterMethodCall, result: @escaping FlutterResult){
+
+        FlyMessenger.sendMeetMessage(messageParams: meetMessageParams){ isSuccess,error,chatMessage in
+            if (isSuccess) {
+                let meetResponse = chatMessage?.toJson()
+//                self.setLastMessage(messageID: chatMessage?.messageId)
+                print("FlyMessenger.sendMeetMessage==**==\(String(describing: meetResponse))")
+                result(meetResponse)
+            }else{
+                //Error Code Checked with SDK
+                if case let .invalid_data(message, _) = error {
+                    result(FlutterError(code: FLErrorCode.INVALID_DATA, message: FLErrorMessage.MESSAGE_SENDING_FAILED, details: message))
+                }else if case let .unexpected(message, code) = error {
+                    if code == ErrorCode.CANNOT_PROCESS{
+                        result(FlutterError(code: FLErrorCode.CANNOT_PROCESS, message: FLErrorMessage.MESSAGE_SENDING_FAILED, details: message))
+                    }else if code == ErrorCode.FORBIDDEN{
+                        result(FlutterError(code: FLErrorCode.FORBIDDEN_ACTION, message: FLErrorMessage.FEATURE_NOT_AVAILABLE, details: message))
+                    }else{
+                        result(FlutterError(code: FLErrorCode.INVALID_DATA, message: FLErrorMessage.MESSAGE_SENDING_FAILED, details: message))
+                    }
+                }else if case let .invalid_jid(message, _) = error {
+                    result(FlutterError(code: FLErrorCode.INVALID_DATA, message: FLErrorMessage.MESSAGE_SENDING_FAILED, details: message))
+                }else{
+                    result(FlutterError(code: FLErrorCode.INVALID_DATA, message: FLErrorMessage.MESSAGE_SENDING_FAILED, details: error?.localizedDescription))
+                }
+            }
+        }
+
+    }
+
     func unFavouriteAllFavouriteMessages(call : FlutterMethodCall, result: @escaping FlutterResult){
         ChatManager.unFavouriteAllFavouriteMessages{(isSuccess, flyError, resultDict) in
             if isSuccess{
@@ -4322,19 +4337,78 @@ let ISEXPORT = true
         
         let barcode = args["barcode"] as? String ?? ""
         
-        WebLoginsManager.shared.handleQrCodeData(qrCodeString: barcode) { isSuccess, message in
+        /// Native side, they are checking for the socket id and then allowing the scanning. So for now we are testing without getting the SocketId
+        WebLoginsManager.shared.getSocketId { isSuccess, message in
+               
             if isSuccess {
-                print("QRCodeScannerViewModel isSuccess")
-                result(isSuccess)
+                DispatchQueue.main.asyncAfter(deadline: .now() + 1) {
+                    WebLoginsManager.shared.handleQrCodeData(qrCodeString: barcode) { isSuccess, message in
+                        if isSuccess {
+                            print("QRCodeScannerViewModel isSuccess")
+                            self.saveWebLoginInfo(qrData: barcode)
+                            result(isSuccess)
+                        }else{
+                            result(FlutterError(code: FLErrorCode.INVALID_DATA, message: FLErrorMessage.QR_LOGIN_FAILED, details: message))
+                        }
+                    }
+                }
+                
             }else{
-                result(FlutterError(code: FLErrorCode.INVALID_DATA, message: FLErrorMessage.QR_LOGIN_FAILED, details: message))
+                print("QRCodeScannerViewModel failed \(message)")
             }
         }
+        
+       
     }
     
+    private func saveWebLoginInfo(qrData : String) {
+        WebLoginsManager.shared.saveWebLogin(qrData: qrData)
+    }
     
     func webLoginDetailsCleared(call : FlutterMethodCall, result: @escaping FlutterResult){
+        WebLoginsManager.shared.reset()
+    }
+    
+    func getWebLoginDetails(call : FlutterMethodCall, result: @escaping FlutterResult){
+        let loginDetails : [WebLoginInfo?] = WebLoginsManager.shared.getWebLogins()
+        print("getWebLoginDetails \(loginDetails)")
         
+        var jsonArray: [[String: Any]] = []
+        
+        let dateFormatter = DateFormatter()
+        dateFormatter.dateFormat = "EEE, d MMM yyyy hh:mm:ss a" // Desired format
+        dateFormatter.locale = Locale(identifier: "en_US") // Ensures proper day/month formatting
+
+        for login in loginDetails.compactMap({ $0 }) {
+            var jsonObject: [String: Any] = [:]
+            
+            jsonObject["id"] = Int.random(in: 0..<100)
+            jsonObject["osName"] = login.platform ?? ""
+            jsonObject["qrUniqeToken"] = login.token ?? ""
+            jsonObject["webBrowserName"] = login.browser ?? ""
+            
+            if let loginTime = login.loginTime {
+                    let timeInSeconds = loginTime / 1_000_000 // Convert microseconds to seconds
+                    let date = Date(timeIntervalSince1970: timeInSeconds) // Create Date object
+                    jsonObject["lastLoginTime"] = dateFormatter.string(from: date) // Format to required string
+                } else {
+                    jsonObject["lastLoginTime"] = ""
+                }
+
+            jsonArray.append(jsonObject)
+        }
+
+        let webLoginJsonArrray = jsonArray.toJSONString()
+        
+        print("login Details json \(webLoginJsonArrray)")
+        
+        result(webLoginJsonArrray)
+        
+    }
+    
+    func logoutWebUser(call : FlutterMethodCall, result: @escaping FlutterResult){
+        WebLoginsManager.shared.logoutFromDevices()
+        result(true)
     }
     
     func sendContactUsInfo(call : FlutterMethodCall, result: @escaping FlutterResult){
@@ -4481,6 +4555,13 @@ let ISEXPORT = true
             result(true)
         }
     }
+    
+    
+    func cancelBackup(call : FlutterMethodCall, result: @escaping FlutterResult) {
+        BackupManager.shared.cancelBackup()
+        
+    }
+